requirements.txt
Django==1.10.5
django-phonenumber-field==1.1.0
easy-thumbnails==2.3
phonenumberslite==8.2.0
Pillow==4.0.0
django-ckeditor
